+++ pkg/apis/certmanager/v1alpha2/types_challenge.go
