+++ pkg/internal/apis/certmanager/v1alpha2/doc.go
@@ -15,9 +15,9 @@ limitations under the License.
 */
 
 // +k8s:conversion-gen=github.com/jetstack/cert-manager/pkg/internal/apis/certmanager
-// +k8s:conversion-gen-external-types=github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1
+// +k8s:conversion-gen-external-types=github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha2
 // +k8s:defaulter-gen=TypeMeta
-// +k8s:defaulter-gen-input=../../../../apis/certmanager/v1alpha1
+// +k8s:defaulter-gen-input=../../../../apis/certmanager/v1alpha2
 
 // +groupName=certmanager.k8s.io
 package v1alpha1
