+++ pkg/internal/apis/certmanager/v1alpha2/zz_generated.defaults.go
